% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_bioguide}
\alias{lookup_bioguide}
\title{Lookup bioguide ID}
\usage{
lookup_bioguide(x, using)
}
\arguments{
\item{x}{Input vector of user IDs}

\item{using}{A data frame lookup dictionary with 'member_user_id' and
'bioguide' columns}
}
\value{
Vector of bioguide IDs (NAs returned for non-matched values)
}
\description{
Converts user IDs into Congressional bioguide IDs.
}
