% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_screen_name}
\alias{lookup_screen_name}
\title{Lookup screen names}
\usage{
lookup_screen_name(x)
}
\arguments{
\item{x}{Input vector of Twitter user IDs}
}
\value{
A vector of screen names (NAs returned for non-matches)
}
\description{
Returns screen names for any user IDs matched to internal users dataset
}
