% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_user_id}
\alias{lookup_user_id}
\title{Lookup user IDs}
\usage{
lookup_user_id(x)
}
\arguments{
\item{x}{Input vector of Twitter screen names}
}
\value{
A vector of user IDs (NAs returned for non-matches)
}
\description{
Returns user IDs for any screen names matched to internal users dataset
}
