% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow-decisions.R
\name{new_follows}
\alias{new_follows}
\title{Count follows}
\usage{
new_follows(x)
}
\arguments{
\item{x}{Friend networks in the form of a recursive list where each element
contains a vector of friends (user IDs)}
}
\value{
Counts represent the number of follow decisions. The first count
will always be missing.
}
\description{
Counts number of follow decisions
}
