% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow-decisions.R
\name{new_unfollows}
\alias{new_unfollows}
\title{Count unfollows}
\usage{
new_unfollows(x)
}
\arguments{
\item{x}{Friend networks in the form of a recursive list where each element
contains a vector of friends (user IDs)}
}
\value{
Counts represent the number of unfollow decisions. The first count
will always be missing.
}
\description{
Counts number of unfollow decisions
}
