% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id2sn.R
\name{sid2sn}
\alias{sid2sn}
\title{Status ID to screen name}
\usage{
sid2sn(x)
}
\arguments{
\item{x}{Input status ID}
}
\value{
Screen name
}
\description{
Converts Twitter status ID to screen name
}
\examples{

## get screen name from status ID
sid2sn("1187362531859259392")

}
