% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow-decisions.R
\name{who_unfollowed}
\alias{who_unfollowed}
\title{Identify unfollowed accounts}
\usage{
who_unfollowed(x)
}
\arguments{
\item{x}{Friend networks in the form of a recursive list where each element
contains a vector of friends (user IDs)}
}
\value{
A list where each element contains the user IDs of accounts
unfollowed (compared to the previous observation).
}
\description{
Returns user identifiers of unfollowed decisions
}
